implement pixel XORing and wrapping
sprites are properly XOR'd to the display, setting the 0xF register as appropriate.
sprites also wrap on the bottom and irght side of the screen. because you can't access negative video memory, wrpaping top to bottom and left to right isn't implemented

+++ chip8.h
@@ -17,6 +17,6 @@
 
 int load_rom();
 void chip8_init();
-void chip8_draw_sprite(int startx, int starty, uint8_t mem, uint8_t size);
+void chip8_draw_sprite(int startx, int starty, uint16_t mem, uint8_t size);
 
 #endif
